added style

+++ src/Map/Vertex.hpp
@@ -1,3 +1,9 @@
+/**
+ * Name: Ya Gao, Qingyang Xu
+ * Emails: yag003@ucsd.edu, q4xu@ucsd.edu
+ * Description: this file contains information for the
+ * Vertex class
+ */
 #ifndef VERTEX_HPP
 #define VERTEX_HPP
 
@@ -14,15 +20,14 @@ class Edge;
 class Vertex {
   public:
     Vertex* parent = nullptr;
-    int dist = INT_MAX;
+    float dist = INT_MAX;
     bool done = false;
+
     const string name;
     float x;
     float y;
     vector<Edge*> outEdges;  // the adjacency list of this vertex that contains
                              // all outgoing edges
-
-    // TODO: you may add more member variables here
 
     /* The constructor that creates a new vertex */
     Vertex(const string& name, float x, float y) : name(name), x(x), y(y) {}
